+++ lib/open_food_network/ga_feature.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module OpenFoodNetwork
+  class GAFeature
+    def initialize(_users); end
+
+    def enabled?(_user)
+      true
+    end
+  end
+end
