+++ lib/open_food_network/null_feature.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module OpenFoodNetwork
+  class NullFeature
+    def enabled?(_user)
+      false
+    end
+  end
+end
